+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 - CodeCaller value builder (moved from ganbarodigital/php-reflection)
+- ExceptionCaller trait
 
 ## 1.0.0 - Sat Jun 27 2015